feat(sqlite): save data to permanent storage database
