+++ oop_challange/Classes/FormulatrixRepositoryManager.cs
@@ -1,19 +1,29 @@
 ﻿using System;
+using System.Data.SQLite;
 
-namespace Oop_Challange
+namespace oop_challange
 {
-	public class FormulatrixRepositoryManager
-	{
-        private Dictionary<string, Tuple<string, int>> dataStore;
-        private bool isInitialized = false;
+    public class FormulatrixRepositoryManager
+    {
+        private SQLiteConnection dbConnection;
+        private const string ConnectionString = "Data Source=../../../database/FormulatrixSQLite.db;Version=3;";
 
         // Initialize the repository for use
         public void Initialize()
         {
-            if (!isInitialized)
+            if (dbConnection == null)
             {
-                dataStore = new Dictionary<string, Tuple<string, int>>();
-                isInitialized = true;
+                dbConnection = new SQLiteConnection(ConnectionString);
+                dbConnection.Open();
+
+                // Create a table if it does not exist
+                string createTableQuery = @"
+                CREATE TABLE IF NOT EXISTS RepositoryItems (
+                    ItemName TEXT PRIMARY KEY,
+                    ItemContent TEXT,
+                    ItemType INTEGER
+                );";
+                ExecuteNonQuery(createTableQuery);
             }
             else
             {
@@ -24,37 +34,49 @@ namespace Oop_Challange
         // Store an item to the repository
         public void Register(string itemName, string itemContent, int itemType)
         {
-            if (!isInitialized)
+            if (dbConnection == null)
             {
                 Console.WriteLine("Repository has not been initialized. Call Initialize method first.");
                 return;
             }
 
-            if (dataStore.ContainsKey(itemName))
+            if (IsItemExists(itemName))
             {
                 Console.WriteLine($"Item with name '{itemName}' already exists. Use a different name.");
                 return;
             }
 
-            // Validate itemContent based on the given itemType
             
-            // Store the item in the dataStore
-            dataStore[itemName] = new Tuple<string, int>(itemContent, itemType);
+            // Store the item in the database
+            string insertQuery = "INSERT INTO RepositoryItems (ItemName, ItemContent, ItemType) VALUES (@ItemName, @ItemContent, @ItemType)";
+            using (var command = new SQLiteCommand(insertQuery, dbConnection))
+            {
+                command.Parameters.AddWithValue("@ItemName", itemName);
+                command.Parameters.AddWithValue("@ItemContent", itemContent);
+                command.Parameters.AddWithValue("@ItemType", itemType);
+                command.ExecuteNonQuery();
+            }
+
             Console.WriteLine($"Item '{itemName}' registered successfully.");
         }
 
         // Retrieve an item from the repository
         public string Retrieve(string itemName)
         {
-            if (!isInitialized)
+            if (dbConnection == null)
             {
                 Console.WriteLine("Repository has not been initialized. Call Initialize method first.");
                 return null;
             }
 
-            if (dataStore.ContainsKey(itemName))
+            if (IsItemExists(itemName))
             {
-                return dataStore[itemName].Item1;
+                string selectQuery = "SELECT ItemContent FROM RepositoryItems WHERE ItemName = @ItemName";
+                using (var command = new SQLiteCommand(selectQuery, dbConnection))
+                {
+                    command.Parameters.AddWithValue("@ItemName", itemName);
+                    return command.ExecuteScalar() as string;
+                }
             }
             else
             {
@@ -66,15 +88,20 @@ namespace Oop_Challange
         // Retrieve the type of the item (JSON or XML)
         public int GetType(string itemName)
         {
-            if (!isInitialized)
+            if (dbConnection == null)
             {
                 Console.WriteLine("Repository has not been initialized. Call Initialize method first.");
                 return -1;
             }
 
-            if (dataStore.ContainsKey(itemName))
+            if (IsItemExists(itemName))
             {
-                return dataStore[itemName].Item2;
+                string selectQuery = "SELECT ItemType FROM RepositoryItems WHERE ItemName = @ItemName";
+                using (var command = new SQLiteCommand(selectQuery, dbConnection))
+                {
+                    command.Parameters.AddWithValue("@ItemName", itemName);
+                    return Convert.ToInt32(command.ExecuteScalar());
+                }
             }
             else
             {
@@ -86,15 +113,21 @@ namespace Oop_Challange
         // Remove an item from the repository
         public void Deregister(string itemName)
         {
-            if (!isInitialized)
+            if (dbConnection == null)
             {
                 Console.WriteLine("Repository has not been initialized. Call Initialize method first.");
                 return;
             }
 
-            if (dataStore.ContainsKey(itemName))
+            if (IsItemExists(itemName))
             {
-                dataStore.Remove(itemName);
+                string deleteQuery = "DELETE FROM RepositoryItems WHERE ItemName = @ItemName";
+                using (var command = new SQLiteCommand(deleteQuery, dbConnection))
+                {
+                    command.Parameters.AddWithValue("@ItemName", itemName);
+                    command.ExecuteNonQuery();
+                }
+
                 Console.WriteLine($"Item '{itemName}' deregistered successfully.");
             }
             else
@@ -102,6 +135,32 @@ namespace Oop_Challange
                 Console.WriteLine($"Item with name '{itemName}' not found.");
             }
         }
+
+        // Check if an item exists in the repository
+        private bool IsItemExists(string itemName)
+        {
+            string selectQuery = "SELECT COUNT(*) FROM RepositoryItems WHERE ItemName = @ItemName";
+            using (var command = new SQLiteCommand(selectQuery, dbConnection))
+            {
+                command.Parameters.AddWithValue("@ItemName", itemName);
+                return Convert.ToInt32(command.ExecuteScalar()) > 0;
+            }
+        }
+
+        // Execute a non-query SQL command
+        private void ExecuteNonQuery(string query)
+        {
+            using (var command = new SQLiteCommand(query, dbConnection))
+            {
+                command.ExecuteNonQuery();
+            }
+        }
+
+        // Close the database connection
+        public void Dispose()
+        {
+            dbConnection?.Close();
+            dbConnection?.Dispose();
+        }
     }
 }
-
