+++ oop_challange/Enums/EnumFileFormat.cs
@@ -8,4 +8,3 @@ namespace oop_challange
         XML
     }
 }
-
